#74 [REFACTOR] 캐싱을 위해 매핑-정렬-페이징 흐름 리팩토링 완료
