+++ src/main/java/healeat/server/web/dto/StoreRequestDto.java
@@ -3,7 +3,6 @@ package healeat.server.web.dto;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class StoreRequestDto {
